==> josh_work/01_cdc_tokenize.ipynb <==
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import gensim

# import nltk
# nltk.download()
# # download punkt

from nltk.tokenize import word_tokenize

df = pd.read_csv('data/train.csv')
df.head(5)

# df['processed'] = df['text'].apply(gensim.utils.simple_preprocess)

df['text_edit'] = df['text']

word_tokenize(df.loc[0,'text_edit'])

df['text_edit'] = df['text'].str.lower()

df['text_edit'] = df['text_edit'].str.replace('\d+','')

import string

string.punctuation

No punctuation

np.sum(df['text_edit'].isin(list(string.punctuation)))

### Create list of tokens to filter out later

df['text_edit'] = df['text_edit'].apply(word_tokenize)

df.head()

Stem words

from nltk.stem.porter import PorterStemmer
porter = PorterStemmer()

def series_stemmer(x):
    return list(map(porter.stem,x))

df['text_edit'] = df['text_edit'].apply(series_stemmer)

df['text_edit'].head(20)

df.loc[14]

df.loc[3,'text_edit']

Remove stop words

# import nltk
# nltk.download()

from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))
print(stop_words)

# len(stop_words)

# print(list(map(porter.stem,stop_words)))

def series_stop_words(x):
    return [w for w in x if not w in stop_words]

%%time
df['text_edit'] = df['text_edit'].apply(series_stop_words)

df['text_edit'].head(20)

Count tokens

from collections import Counter

token_cnt = Counter()

for _,x in df['text_edit'].iteritems():
    token_cnt.update(x)

token_cnt.most_common()

tokenseries = pd.Series(token_cnt)

tokenseries = tokenseries.sort_values(ascending=False)

Words to remove

pd.options.display.max_colwidth = 70

df.loc[df['text'].str.contains(' FB '),'text']

to_remove = {'at','a' ,'onto', 'state', 
       'yf', 'm', 'f', 'to', 'on', 'male', 'female',
             'yom', 'yof', 'yom','pt', 'yo','ym',
       'rt',   'lt', 'o','l', 'r','left', 'right',
        }

tokenseries.iloc[:50].index

tokenseries.iloc[:50]

# to_remove = to_remove.union(set(tokenseries[tokenseries <= 5].index))

def series_remove_words(x):
    return [w for w in x if not w in to_remove]

%%time
df['text_edit'] = df['text_edit'].apply(series_remove_words)

df['text_edit'].head(20)

df['text_edit'].apply(len).value_counts().sort_index()

df['token_len'] = df['text_edit'].apply(len)

df[df['token_len']==30]

df.head()

docs = df['text_edit']

%%time
mod = gensim.models.Word2Vec(
        list(docs),
        size=150,
        window=7,
        min_count=3,
        iter=10,
        sg=1)

w1 = 'arm'
mod.wv.most_similar(w1)

w1 = 'lac'
mod.wv.most_similar(w1)

w1 = 'dx'
mod.wv.most_similar(w1)

w1 = 'fx'
mod.wv.most_similar(w1)

w1 = 'clipper'
mod.wv.most_similar(w1)

w1 = 'twsite'
mod.wv.most_similar(w1)

w1 = 'lacerat'
mod.wv.most_similar(w1)

Most common words

w1 = 'work'
mod.wv.most_similar(w1)

for i in range(50):
    w1 = mod.wv.index2word[i]
    print(f"\n-----{w1}---------------------------------")
    topw = mod.wv.most_similar(w1,topn=3)
    topw = [w[0] for w in topw]
    print(topw)

Least Common

vocab_size = len(mod.wv.vocab)
mod.wv.index2word[vocab_size - 10:]

w1 = 'handsaw'
mod.wv.most_similar(w1)

w1 = 'paindu'
mod.wv.most_similar(w1)

df.to_parquet('data/tokenized.pqt')

df.to_pickle('data/tokenized.pkl')
