==> josh_work/03_fit_Keras_model-Copy1.ipynb <==
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import gensim

import tensorflow as tf

config = tf.ConfigProto()
config.gpu_options.allow_growth = True
tf.keras.backend.set_session(tf.Session(config=config))

# import os
# os.environ['CUDA_VISIBLE_DEVICES'] = '-1'

# from tensorflow.keras.preprocessing.text import Tokenizer
# from tensorflow.keras.preprocessing.sequence import pad_sequences
# from tensorflow.keras.models import Sequential
# from tensorflow.keras.layers import Dense
# from tensorflow.keras.layers import Flatten
# from tensorflow.keras.layers import Embedding
# from tensorflow.keras.layers import Conv1D
# from tensorflow.keras.layers import MaxPooling1D
# from tensorflow.keras import optimizers
# from keras.utils import np_utils
# from sklearn.preprocessing import LabelEncoder

from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.models import Sequential
from keras.layers import Dense
from keras.layers import Flatten
from keras.layers import Embedding
from keras.layers.convolutional import Conv1D
from keras.layers.convolutional import MaxPooling1D
from sklearn.preprocessing import LabelEncoder
from keras.utils import np_utils

from keras import optimizers

df = pd.read_pickle('./data/tokenized.pkl')
df['text_edit'] = df['text_edit'].apply(list)
df.head(5)

Create reference vocabulary used for training

# vocab = set()
# for _,e in df['text_edit'].iteritems():
#     vocab = vocab.union(set(e))
#
# len(vocab)

# len(vocab)

### Try Keras that fits the embedding

Determine maximum token sequence length for training. We don't want to simply use that max as reviews 20 or longer are rare. Based on distribution below we select a max length of 15.

df['token_len'].value_counts().sort_index()

max_len = 15

docs = list(df['text_edit'].str[:max_len])

tokenizer = Tokenizer()

tokenizer.fit_on_texts(docs)

encoded_doc = tokenizer.texts_to_sequences(docs)

encoded_doc[:3]

xtrain = pad_sequences(encoded_doc, maxlen=max_len, padding='post')

xtrain[0]

vocab_size = len(tokenizer.word_index) + 1
# we add +1 to account for missing words

Create target matrix.  
As this is a multi-class model keras requires the target to be in the form of a matrix.

encoder = LabelEncoder()

encoder.fit(df['event'].values)

yencoded = encoder.fit_transform(df['event'].values)

encoder.inverse_transform([42])

ytrain = np_utils.to_categorical(yencoded)

target_size = len(np.unique(yencoded))
target_size

Define Model

# model = Sequential()
# model.add(Embedding(vocab_size, 1, input_length=max_len))
# model.add(Conv1D(filters=1, kernel_size=2, activation='relu'))
# model.add(MaxPooling1D(pool_size=2))
# model.add(Flatten())
# model.add(Dense(1, activation='relu'))
# model.add(Dense(target_size, activation='softmax'))

model = Sequential()
model.add(Embedding(vocab_size, 100, input_length=max_len))
model.add(Conv1D(filters=32, kernel_size=8, activation='relu'))
model.add(MaxPooling1D(pool_size=2))
model.add(Flatten())
model.add(Dense(100, activation='relu'))
model.add(Dense(target_size, activation='softmax'))

print(model.summary())

model.compile(loss='categorical_crossentropy', optimizer=optimizers.Adam(lr=0.0001), metrics=['accuracy'])

model.fit(xtrain, ytrain, epochs=10, verbose=1, validation_split=.2, batch_size=32)

model.fit(xtrain, ytrain, epochs=10, verbose=1, validation_split=.2, batch_size=16)

model.compile(loss='categorical_crossentropy', optimizer=optimizers.Adam(lr=0.001), metrics=['accuracy'])

model.fit(xtrain, ytrain, epochs=10, verbose=1, validation_split=.2, batch_size=128)

model.compile(loss='categorical_crossentropy', optimizer=optimizers.adam(lr=0.0005), metrics=['accuracy'])

model.fit(xtrain, ytrain, epochs=2,verbose=1,validation_split=.2)

model.compile(loss='categorical_crossentropy', optimizer=optimizers.adam(lr=0.0001), metrics=['accuracy'])

model.fit(xtrain, ytrain, epochs=5,verbose=1,validation_split=.2, batch_size=256)

ytrain.argmax(1)

yprob = model.predict(xtrain)

yprob[0].argmax()

ypred = yprob.argmax(1)

dfres = pd.DataFrame(data=np.column_stack([ytrain.argmax(1),ypred]),columns=['act','pred'])
dfres.head()

# pd.crosstab(dfres['act'],dfres['pred']).to_excel('./results/crosstab_accuracy.xlsx')

np.sum(dfres['act'] == dfres['pred'])/dfres.shape[0]

pd.concat([dfres['act'].value_counts(),dfres['pred'].value_counts()],1).sort_index().fillna('')


### Fit on all data instead of using a validation set
Based on above fit we will fit 10 epochs

model = Sequential()
model.add(Embedding(vocab_size, 100, input_length=max_len))
model.add(Conv1D(filters=32, kernel_size=8, activation='relu'))
model.add(MaxPooling1D(pool_size=2))
model.add(Flatten())
model.add(Dense(100, activation='relu'))
model.add(Dense(target_size, activation='softmax'))

print(model.summary())

model.compile(loss='categorical_crossentropy', optimizer=optimizers.adam(lr=0.0001), metrics=['accuracy'])

model.fit(xtrain, ytrain, epochs=10,verbose=1)

model.save('./models/keras_try1.h5')
