==> josh_work/05_keywords.ipynb <==
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from event_keywords import event_keywords

df = pd.read_csv('./data/train.csv')
df.head(5)

Create keyword indicators

events = [k for k in event_keywords.keys() if len(event_keywords[k]) > 0]
print(events)

%%time
for event in events:
    df[event+'_ind'] = df['text'].apply(lambda x: 1*any([k in x for k in event_keywords[event]]))

event_indicators = [e+'_ind' for e in events]

df.head()

df[event_indicators].sum(0)

df[event_indicators].to_pickle('data/event_indicators.pkl')
