==> josh_work/06_lightgbm.ipynb <==
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import lightgbm as lgb
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import accuracy_score
from event_keywords import event_keywords
from hyperopt import hp, tpe, fmin, Trials

Read in processed data sources and join

df = pd.read_pickle('./data/tokenized.pkl')
df.head(5)

events = sorted(list(df['event'].drop_duplicates()))

df_keras = pd.read_pickle('data/OOS_keras_preds.pkl')
df_keras.head()

prob_cols = [str(e)+'_prob' for e in events]
pred_cols = [str(e)+'_pred' for e in events]

df_keywords = pd.read_pickle('data/event_indicators.pkl')
df_keywords.head()

event_cols = [k+'_ind' for k in event_keywords.keys() if len(event_keywords[k]) > 0]
print(event_cols)

df.columns.values

dfx = pd.concat([df[['sex', 'age','token_len']],df_keywords,df_keras],axis=1)
print(dfx.shape)
dfx.head()

dfy = df[['event']]

del df,df_keywords,df_keras

### Lightgbm model

Use same multi-class target format we used with keras

encoder = LabelEncoder()

encoder.fit(dfy.values.ravel())

yencoded = encoder.transform(dfy.values.ravel())

Create train validation split

x_train, x_test, y_train, y_test = train_test_split(dfx,yencoded,
                                                    test_size=0.2,random_state=42)

x_train, x_test, y_train, y_test = train_test_split(dfx,yencoded,
                                                    test_size=0.2,random_state=1)

lgbtrain = lgb.Dataset(x_train, y_train)
lgbeval = lgb.Dataset(x_test, y_test, reference=lgbtrain)

params = {
'objective': 'multiclass',
'metric': {'multi_logloss','multi_error'},
'num_class':48,
'num_leaves': 31,
'learning_rate': 0.01,
'feature_fraction': 0.9,
'bagging_fraction': 0.7,
'bagging_freq': 1,
# 'verbose': 1,
}

params = {
'objective': 'multiclass',
'metric': {'multi_error'},
'num_class':48,
'num_leaves': 31,
# 'num_leaves': 124,
'learning_rate': 0.01,
'feature_fraction': 0.8,
'bagging_fraction': 0.5,
'bagging_freq': 1,
# 'verbose': 1,
}

lgbm = lgb.train(params,lgbtrain,num_boost_round=2000,
                 valid_sets=[lgbtrain,lgbeval],verbose_eval=50,early_stopping_rounds=100)

accuracy_score(y_test,lgbm.predict(x_test).argmax(1))

### Hyperopt

dfx.columns.values

train_cols = ['sex', 'age', 'token_len'] + prob_cols + event_cols
x_train, x_test, y_train, y_test = train_test_split(dfx[train_cols],yencoded,
                                                    test_size=0.2,random_state=42)

lgbtrain = lgb.Dataset(x_train, y_train)
lgbeval = lgb.Dataset(x_test, y_test, reference=lgbtrain)

params = {
    'objective': 'multiclass',
    'metric': {'multi_error'},
    'num_class':48,
    'num_leaves': 31,
    'learning_rate': 0.01,
    'feature_fraction': 0.8,
    'bagging_fraction': 0.5,
    'bagging_freq': 1,
    'verbose':-1
    }

def objective(args):
    # param updates
    params['num_leaves'] = args['num_leaves']
    params['learning_rate'] = args['learning_rate']
    params['feature_fraction'] = args['feature_fraction']
    params['bagging_fraction'] = args['bagging_fraction']
    params['min_data_in_leaf'] = args['min_data_in_leaf']
    lgbm = lgb.train(params,lgbtrain,
                     num_boost_round=500,
                 valid_sets=lgbeval,
                     verbose_eval=-1,
                     early_stopping_rounds=20)
    return 1 - accuracy_score(y_test,lgbm.predict(x_test).argmax(1))

space = dict([('learning_rate', hp.loguniform('learning_rate', 
                                                     np.log(0.0005), np.log(0.05))),
                    ('num_leaves', hp.choice('num_leaves', range(10, 100, 10))),
                    ('min_data_in_leaf', hp.choice('min_data_in_leaf', 
                                                  range(100, 1000, 100))),
                    ('feature_fraction', hp.uniform('feature_fraction', 0.5, 1.0)),
                    ('bagging_fraction', hp.uniform('subsample', 0.5, 1.0))
                    ])

trials = Trials()

best = fmin(objective, space, trials=trials, algo=tpe.suggest, max_evals=100)

trials.results

trials.idxs_vals

tpe_results = pd.DataFrame({'loss': np.round([x['loss'] for x in trials.results],3), 
                            'feature_fraction': np.round(trials.idxs_vals[1]['feature_fraction'],2),
                            'learning_rate': np.round(trials.idxs_vals[1]['learning_rate'],4),
                            'min_data_in_leaf': trials.idxs_vals[1]['min_data_in_leaf'],
                            'num_leaves': trials.idxs_vals[1]['num_leaves'],
                            'subsample': np.round(trials.idxs_vals[1]['subsample'],2),
                           })

tpe_results

trials.idxs_vals[1]

sp = hp.uniform('x',0,10)

import hyperopt

hyperopt.pyll.stochastic.sample(sp)
